feat: add signup and logout system, rework auth to work with refresh tokens
Auth now works with refresh tokens, where an access token is granted for
a short amount of time (here 15 minutes), which is verified with jwt to
grant access. If the token expires, refresh it with the longer lived
refresh token (here the age is 30 days), along with the refresh token to
keep active users logged in. Logging out will invalidate the user's
currently active refresh token by deleting it from the database.