feat: reworked auth system to use refresh tokens, added admin privilege checking, signing up, and logging out (#12)
* feat: add signup and logout system, rework auth to work with refresh tokens

Auth now works with refresh tokens, where an access token is granted for
a short amount of time (here 15 minutes), which is verified with jwt to
grant access. If the token expires, refresh it with the longer lived
refresh token (here the age is 30 days), along with the refresh token to
keep active users logged in. Logging out will invalidate the user's
currently active refresh token by deleting it from the database.

* feat: add `is_admin` column to users table to identify admins and add admin privilege checking middleware

Access tokens will now be signed with the information about whether a
user is an admin or not, and this gets checked in middleware